Refactor for norm compliance
